+++ lib/themes/dark_theme.dart
@@ -11,14 +11,20 @@ ThemeData darkTheme = ThemeData(
   cardColor: Nord.nord3,
   toggleableActiveColor: Nord.nord11,
   appBarTheme: AppBarTheme(
-    shadowColor: Colors.transparent,
-    systemOverlayStyle: SystemUiOverlayStyle(
-        statusBarColor: Colors.transparent,
-        systemNavigationBarColor: Nord.nord0,
-        systemNavigationBarIconBrightness: Brightness.light),
-  ),
+      brightness: Brightness.dark,
+      color: Nord.nord1,
+      shadowColor: Colors.transparent,
+      textTheme: TextTheme(
+        headline6: TextStyle(color: Nord.nord6, fontSize: 22),
+      ),
+      iconTheme: IconThemeData(color: Nord.nord6)),
   bottomNavigationBarTheme: BottomNavigationBarThemeData(
       backgroundColor: Nord.nord0,
       selectedItemColor: Nord.nord11,
       unselectedItemColor: Nord.nord6),
 );
+
+SystemUiOverlayStyle darkSystemUiOverlayStyle = SystemUiOverlayStyle(
+    statusBarColor: Colors.transparent,
+    systemNavigationBarColor: Nord.nord0,
+    systemNavigationBarIconBrightness: Brightness.light);